+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add `after_running_test_group` and `before_running_test_group` events
 - Add `Petitest::TestGroup#nest_level`
 - Add `Petitest::TestGroup#sub_test_group`
+- Add document style report
 - Improve error report format
 - Improve default `#assert` message